prevent multiple clients initialization
